Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4
python-dotenv
python-multipart
python-slugify
httpx
